+++ cmd/dmsgweb/commands/root.go
@@ -1,0 +1,351 @@
+// Package commands cmd/dmsgweb/commands/dmsgweb.go
+package commands
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"net"
+	"net/http"
+	"os"
+	"runtime"
+	"strconv"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/bitfield/script"
+	"github.com/gin-gonic/gin"
+	"github.com/skycoin/skywire-utilities/pkg/cipher"
+	"github.com/skycoin/skywire-utilities/pkg/logging"
+
+	"github.com/skycoin/dmsg/pkg/disc"
+	dmsg "github.com/skycoin/dmsg/pkg/dmsg"
+)
+
+
+var (
+	httpC              http.Client
+	dmsgDisc           string
+	dmsgSessions       int
+	filterDomainSuffix string
+	sk                 cipher.SecKey
+	pk                 cipher.PubKey
+	dmsgWebLog         *logging.Logger
+	logLvl             string
+	webPort            uint
+	proxyPort          uint
+	addProxy           string
+	resolveDmsgAddr    string
+	wg                 sync.WaitGroup
+	isEnvs             bool
+	dmsgPort   []uint
+	dmsgSess   int
+	wl         string
+	wlkeys     []cipher.PubKey
+	localPort  []uint
+	err        error
+)
+
+// Execute executes root CLI command.
+func Execute() {
+	if err := RootCmd.Execute(); err != nil {
+		log.Fatal("Failed to execute command: ", err)
+	}
+}
+
+func startDmsg(ctx context.Context, pk cipher.PubKey, sk cipher.SecKey) (dmsgC *dmsg.Client, stop func(), err error) {
+	dmsgC = dmsg.NewClient(pk, sk, disc.NewHTTP(dmsgDisc, &http.Client{}, dmsgWebLog), &dmsg.Config{MinSessions: dmsgSessions})
+	go dmsgC.Serve(context.Background())
+
+	stop = func() {
+		err := dmsgC.Close()
+		dmsgWebLog.WithError(err).Debug("Disconnected from dmsg network.")
+		fmt.Printf("\n")
+	}
+	dmsgWebLog.WithField("public_key", pk.String()).WithField("dmsg_disc", dmsgDisc).
+		Debug("Connecting to dmsg network...")
+
+	select {
+	case <-ctx.Done():
+		stop()
+		os.Exit(0)
+		return nil, nil, ctx.Err()
+
+	case <-dmsgC.Ready():
+		dmsgWebLog.Debug("Dmsg network ready.")
+		return dmsgC, stop, nil
+	}
+}
+
+func scriptExecString(s, envfile string) string {
+	if runtime.GOOS == "windows" {
+		var variable, defaultvalue string
+		if strings.Contains(s, ":-") {
+			parts := strings.SplitN(s, ":-", 2)
+			variable = parts[0] + "}"
+			defaultvalue = strings.TrimRight(parts[1], "}")
+		} else {
+			variable = s
+			defaultvalue = ""
+		}
+		out, err := script.Exec(fmt.Sprintf(`powershell -c '$SKYENV = "%s"; if ($SKYENV -ne "" -and (Test-Path $SKYENV)) { . $SKYENV }; echo %s"`, envfile, variable)).String()
+		if err == nil {
+			if (out == "") || (out == variable) {
+				return defaultvalue
+			}
+			return strings.TrimRight(out, "\n")
+		}
+		return defaultvalue
+	}
+	z, err := script.Exec(fmt.Sprintf(`sh -c 'SKYENV=%s ; if [[ $SKYENV != "" ]] && [[ -f $SKYENV ]] ; then source $SKYENV ; fi ; printf "%s"'`, envfile, s)).String()
+	if err == nil {
+		return strings.TrimSpace(z)
+	}
+	return ""
+}
+
+func scriptExecArray(s, envfile string) string {
+	if runtime.GOOS == "windows" {
+		variable := s
+		if strings.Contains(variable, "[@]}") {
+			variable = strings.TrimRight(variable, "[@]}")
+			variable = strings.TrimRight(variable, "{")
+		}
+		out, err := script.Exec(fmt.Sprintf(`powershell -c '$SKYENV = "%s"; if ($SKYENV -ne "" -and (Test-Path $SKYENV)) { . $SKYENV }; foreach ($item in %s) { Write-Host $item }'`, envfile, variable)).Slice()
+		if err == nil {
+			if len(out) != 0 {
+				return ""
+			}
+			return strings.Join(out, ",")
+		}
+	}
+	y, err := script.Exec(fmt.Sprintf(`bash -c 'SKYENV=%s ; if [[ $SKYENV != "" ]] && [[ -f $SKYENV ]] ; then source $SKYENV ; fi ; for _i in %s ; do echo "$_i" ; done'`, envfile, s)).Slice()
+	if err == nil {
+		return strings.Join(y, ",")
+	}
+	return ""
+}
+
+func scriptExecUintSlice(s, envfile string) []uint {
+	var out []string
+	var err error
+
+	if runtime.GOOS == "windows" {
+		variable := s
+		if strings.Contains(variable, "[@]}") {
+			variable = strings.TrimRight(variable, "[@]}")
+			variable = strings.TrimRight(variable, "{")
+		}
+		out, err = script.Exec(fmt.Sprintf(`powershell -c '$SKYENV = "%s"; if ($SKYENV -ne "" -and (Test-Path $SKYENV)) { . $SKYENV }; foreach ($item in %s) { Write-Host $item }'`, envfile, variable)).Slice()
+	} else {
+		out, err = script.Exec(fmt.Sprintf(`bash -c 'SKYENV=%s ; if [[ $SKYENV != "" ]] && [[ -f $SKYENV ]] ; then source $SKYENV ; fi ; for _i in %s ; do echo "$_i" ; done'`, envfile, s)).Slice()
+	}
+
+	if err != nil {
+		return []uint{}
+	}
+
+	var res []uint
+	for _, item := range out {
+		num, err := strconv.ParseUint(item, 10, 64)
+		if err == nil {
+			res = append(res, uint(num))
+		}
+	}
+
+	return res
+}
+
+func scriptExecInt(s, envfile string) int {
+	if runtime.GOOS == "windows" {
+		var variable string
+		if strings.Contains(s, ":-") {
+			parts := strings.SplitN(s, ":-", 2)
+			variable = parts[0] + "}"
+		} else {
+			variable = s
+		}
+		out, err := script.Exec(fmt.Sprintf(`powershell -c '$SKYENV = "%s"; if ($SKYENV -ne "" -and (Test-Path $SKYENV)) { . $SKYENV }; echo %s"`, envfile, variable)).String()
+		if err == nil {
+			if (out == "") || (out == variable) {
+				return 0
+			}
+			i, err := strconv.Atoi(strings.TrimSpace(strings.TrimRight(out, "\n")))
+			if err == nil {
+				return i
+			}
+			return 0
+		}
+		return 0
+	}
+	z, err := script.Exec(fmt.Sprintf(`sh -c 'SKYENV=%s ; if [[ $SKYENV != "" ]] && [[ -f $SKYENV ]] ; then source $SKYENV ; fi ; printf "%s"'`, envfile, s)).String()
+	if err == nil {
+		if z == "" {
+			return 0
+		}
+		i, err := strconv.Atoi(z)
+		if err == nil {
+			return i
+		}
+	}
+	return 0
+}
+func scriptExecUint(s, envfile string) uint {
+	if runtime.GOOS == "windows" {
+		var variable string
+		if strings.Contains(s, ":-") {
+			parts := strings.SplitN(s, ":-", 2)
+			variable = parts[0] + "}"
+		} else {
+			variable = s
+		}
+		out, err := script.Exec(fmt.Sprintf(`powershell -c '$SKYENV = "%s"; if ($SKYENV -ne "" -and (Test-Path $SKYENV)) { . $SKYENV }; echo %s"`, envfile, variable)).String()
+		if err == nil {
+			if (out == "") || (out == variable) {
+				return 0
+			}
+			i, err := strconv.Atoi(strings.TrimSpace(strings.TrimRight(out, "\n")))
+			if err == nil {
+				return uint(i)
+			}
+			return 0
+		}
+		return 0
+	}
+	z, err := script.Exec(fmt.Sprintf(`sh -c 'SKYENV=%s ; if [[ $SKYENV != "" ]] && [[ -f $SKYENV ]] ; then source $SKYENV ; fi ; printf "%s"'`, envfile, s)).String()
+	if err == nil {
+		if z == "" {
+			return 0
+		}
+		i, err := strconv.Atoi(z)
+		if err == nil {
+			return uint(i)
+		}
+	}
+	return uint(0)
+}
+
+
+
+
+func whitelistAuth(whitelistedPKs []cipher.PubKey) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		remotePK, _, err := net.SplitHostPort(c.Request.RemoteAddr)
+		if err != nil {
+			c.Writer.WriteHeader(http.StatusInternalServerError)
+			c.Writer.Write([]byte("500 Internal Server Error")) //nolint
+			c.AbortWithStatus(http.StatusInternalServerError)
+			return
+		}
+		whitelisted := false
+		if len(whitelistedPKs) == 0 {
+			whitelisted = true
+		} else {
+			for _, whitelistedPK := range whitelistedPKs {
+				if remotePK == whitelistedPK.String() {
+					whitelisted = true
+					break
+				}
+			}
+		}
+		if whitelisted {
+			c.Next()
+		} else {
+			c.Writer.WriteHeader(http.StatusUnauthorized)
+			c.Writer.Write([]byte("401 Unauthorized")) //nolint
+			c.AbortWithStatus(http.StatusUnauthorized)
+			return
+		}
+	}
+}
+
+type ginHandler struct {
+	Router *gin.Engine
+}
+
+func (h *ginHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	h.Router.ServeHTTP(w, r)
+}
+
+
+func loggingMiddleware() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		start := time.Now()
+		c.Next()
+		latency := time.Since(start)
+		if latency > time.Minute {
+			latency = latency.Truncate(time.Second)
+		}
+		statusCode := c.Writer.Status()
+		method := c.Request.Method
+		path := c.Request.URL.Path
+		// Get the background color based on the status code
+		statusCodeBackgroundColor := getBackgroundColor(statusCode)
+		// Get the method color
+		methodColor := getMethodColor(method)
+		// Print the logging in a custom format which includes the publickeyfrom c.Request.RemoteAddr ex.:
+		// [DMSGHTTP] 2023/05/18 - 19:43:15 | 200 |    10.80885ms |                 | 02b5ee5333aa6b7f5fc623b7d5f35f505cb7f974e98a70751cf41962f84c8c4637:49153 | GET      /node-info.json
+		fmt.Printf("[DMSGWEB] %s |%s %3d %s| %13v | %15s | %72s |%s %-7s %s %s\n",
+			time.Now().Format("2006/01/02 - 15:04:05"),
+			statusCodeBackgroundColor,
+			statusCode,
+			resetColor(),
+			latency,
+			c.ClientIP(),
+			c.Request.RemoteAddr,
+			methodColor,
+			method,
+			resetColor(),
+			path,
+		)
+	}
+}
+func getBackgroundColor(statusCode int) string {
+	switch {
+	case statusCode >= http.StatusOK && statusCode < http.StatusMultipleChoices:
+		return green
+	case statusCode >= http.StatusMultipleChoices && statusCode < http.StatusBadRequest:
+		return white
+	case statusCode >= http.StatusBadRequest && statusCode < http.StatusInternalServerError:
+		return yellow
+	default:
+		return red
+	}
+}
+
+func getMethodColor(method string) string {
+	switch method {
+	case http.MethodGet:
+		return blue
+	case http.MethodPost:
+		return cyan
+	case http.MethodPut:
+		return yellow
+	case http.MethodDelete:
+		return red
+	case http.MethodPatch:
+		return green
+	case http.MethodHead:
+		return magenta
+	case http.MethodOptions:
+		return white
+	default:
+		return reset
+	}
+}
+
+func resetColor() string {
+	return reset
+}
+
+const (
+	green   = "\033[97;42m"
+	white   = "\033[90;47m"
+	yellow  = "\033[90;43m"
+	red     = "\033[97;41m"
+	blue    = "\033[97;44m"
+	magenta = "\033[97;45m"
+	cyan    = "\033[97;46m"
+	reset   = "\033[0m"
+)
